implement on 1.8.9
- fix a few bugs
- update changelog

+++ 1.21.5/src/main/java/io/github/axolotlclient/modules/zoom/Zoom.java
@@ -165,6 +165,6 @@ public class Zoom extends AbstractModule {
 
 	public void tick() {
 		lastAnimatedFactor = animatedFactor;
-		animatedFactor += (float) ((targetFactor - animatedFactor) * (zoomSpeed.get() / 10F));
+		animatedFactor += (targetFactor - animatedFactor) * (zoomSpeed.get() / 10F);
 	}
 }
